back-end/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
pyjwt
